requirements.txt
google-cloud-speech ~= 1.3.1
